Moving ThemesDictionary and ControlsDictionary to parent namespace
